Add files via upload
Naruto: Classic is my topic for creating this website. I've used HTML along with CSS and JavaScript languages to design this site. This site is merely a summarizing and more like a review for this anime. I've almost covered half of the former language and a bit of the later.